chore(rig-gemini-grpc): move inline test modules to sibling files
